backend-python/requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
pydantic==2.12.5
qdrant-client==1.16.2
python-dotenv==1.2.1
requests==2.32.5